+++ R/grid_code_paths.R
@@ -35,6 +35,7 @@ tune_grid_loop <- function(resamples, grid, workflow, metrics, control, rng) {
           ns = "tune"
         )
 
+        # Likely want to debug with `debugonce(tune_grid_loop_iter)`
         tune_grid_loop_iter_safely(
           split = split,
           grid_info = grid_info,
@@ -70,6 +71,7 @@ tune_grid_loop <- function(resamples, grid, workflow, metrics, control, rng) {
 
           grid_info_row <- vctrs::vec_slice(grid_info, row)
 
+          # Likely want to debug with `debugonce(tune_grid_loop_iter)`
           tune_grid_loop_iter_safely(
             split = split,
             grid_info = grid_info_row,
@@ -131,7 +133,7 @@ tune_grid_loop_iter <- function(split,
                                 control,
                                 seed) {
   load_pkgs(workflow)
-  load_namespace(control$pkgs)
+  .load_namespace(control$pkgs)
 
   # After package loading to avoid potential package RNG manipulation
   if (!is.null(seed)) {
@@ -211,7 +213,7 @@ tune_grid_loop_iter <- function(split,
       grid_preprocessor = iter_grid_preprocessor
     )
 
-    workflow <- catch_and_log(
+    workflow <- .catch_and_log(
       .expr = .fit_pre(workflow, training),
       control,
       split,
@@ -250,8 +252,8 @@ tune_grid_loop_iter <- function(split,
 
       workflow <- finalize_workflow_spec(workflow, iter_grid_model)
 
-      workflow <- catch_and_log_fit(
-        expr = .fit_model(workflow, control_workflow),
+      workflow <- .catch_and_log_fit(
+        .expr = .fit_model(workflow, control_workflow),
         control,
         split,
         iter_msg_model,
@@ -299,7 +301,7 @@ tune_grid_loop_iter <- function(split,
 
       iter_msg_predictions <- paste(iter_msg_model, "(predictions)")
 
-      iter_predictions <- catch_and_log(
+      iter_predictions <- .catch_and_log(
         predict_model(split, workflow, iter_grid, metrics, iter_submodels),
         control,
         split,
@@ -355,6 +357,7 @@ tune_grid_loop_iter_safely <- function(split,
                                        seed) {
   tune_grid_loop_iter_wrapper <- super_safely(tune_grid_loop_iter)
 
+  # Likely want to debug with `debugonce(tune_grid_loop_iter)`
   result <- tune_grid_loop_iter_wrapper(
     split,
     grid_info,